requirements.txt
flask
gunicorn
openai
unstructured
pdf2image
pytesseract
pillow
python-dotenv
